fix: monir text fixes
