--- cv/jsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es6",
-    "module": "commonjs"
-  },
-  "exclude": ["node_modules", "dist"],
-  "include": ["src/**/*"]
-}
