--- cv/package.json
@@ -1,16 +1,0 @@
-{
-  "version": "0.1.0",
-  "description": "My new js application",
-  "private": true,
-  "name": "cv",
-  "scripts": {
-    "start": "rss-scripts start",
-    "build": "rss-scripts build"
-  },
-  "devDependencies": {
-    "rss-scripts": "^0.2.4"
-  },
-  "dependencies": {
-    "highlight.js": "^11.3.1"
-  }
-}
