--- cv/src/style.css
@@ -1,69 +1,0 @@
-
-*,
-*::before,
-*::after {
-    -webkit-box-sizing: border-box;
-    box-sizing: border-box;
-}
-
-ul[class],
-ol[class] {
-    padding: 0;
-}
-
-body,
-h1,
-h2,
-h3,
-h4,
-h5,
-p,
-ul,
-ol,
-li,
-figure,
-figcaption,
-blockquote,
-dl,
-dd {
-    margin: 0;
-}
-
-p {
-    margin-bottom: .3rem;
-}
-
-html, body {
-    height: 100%;
-    overflow: auto;
-}
-
-html {
-    font-size: 15px;
-    scroll-behavior: smooth;
-    height: auto;
-    transition: font-size .5s ease
-}
-
-body {
-    font-family: 'Open Sans', sans-serif;
-    font-size: 16px;
-    min-height: 100vh;
-    text-rendering: optimizeSpeed;
-    line-height: 1.6;
-    background-color: #d4d8d7;
-    color: #3B3A3B;
-    position: relative;
-    min-width: 320px;
-}
-
-ul[class],
-ol[class] {
-    list-style: none;
-}
-
-img {
-    max-width: 100%;
-    display: block;
-}
-
